requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
alembic==1.13.2